# src/main/java/io/swagger/api/HelloWorldApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (2.3.1).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package io.swagger.api;

import io.swagger.model.HelloWorld;
import io.swagger.model.HelloWorldRequest;
import io.swagger.annotations.*;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestHeader;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RequestPart;
import org.springframework.web.multipart.MultipartFile;

import javax.validation.Valid;
import javax.validation.constraints.*;
import java.util.List;
@javax.annotation.Generated(value = "io.swagger.codegen.languages.SpringCodegen", date = "2018-09-07T07:48:13.195Z")

@Api(value = "helloworld", description = "the helloworld API")
public interface HelloWorldApi {

    @ApiOperation(value = "Get HelloWorld.", nickname = "getHelloWorld", notes = "Get simple Hello World greeting.", response = HelloWorld.class, tags={ "HelloWorld", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Hello World Message", response = HelloWorld.class) })
    @RequestMapping(value = "/helloworld",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<HelloWorld> getHelloWorld();


    @ApiOperation(value = "Get HelloWorld for my friends.", nickname = "getHelloWorldForMyFriends", notes = "Get simple Hello World greeting with your name for your friends.", response = HelloWorld.class, responseContainer = "List", tags={ "HelloWorld", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Hello World Message", response = HelloWorld.class, responseContainer = "List") })
    @RequestMapping(value = "/helloworld",
        produces = { "application/json" }, 
        consumes = { "application/json" },
        method = RequestMethod.POST)
    ResponseEntity<List<HelloWorld>> getHelloWorldForMyFriends(@ApiParam(value = "HelloWorld request" ,required=true )  @Valid @RequestBody HelloWorldRequest body);


    @ApiOperation(value = "Get HelloWorld form me.", nickname = "getHelloWorldWithName", notes = "Get simple Hello World greeting with your name.", response = HelloWorld.class, tags={ "HelloWorld", })
    @ApiResponses(value = { 
        @ApiResponse(code = 200, message = "Hello World Message", response = HelloWorld.class) })
    @RequestMapping(value = "/helloworld/{name}",
        produces = { "application/json" }, 
        method = RequestMethod.GET)
    ResponseEntity<HelloWorld> getHelloWorldWithName(@ApiParam(value = "Enter your name",required=true) @PathVariable("name") String name);

}